test(ui): narrow work chart card contracts
